+++ app/models/home_page.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class HomePage < ApplicationRecord
+  validates :sections, presence: true
+end